types edit

+++ TestLibrary/Figure.cs
@@ -8,8 +8,7 @@ namespace TestLibrary
 {
     public abstract class Figure
     {
-        public abstract double calculateArea();
+        public abstract decimal calculateArea();
         
-
     }
 }
